+++ src/main/java/aztech/modern_industrialization/material/MIMaterialSetup.java
@@ -24,8 +24,8 @@
 package aztech.modern_industrialization.material;
 
 import aztech.modern_industrialization.MIIdentifier;
-import java.util.Map;
-import java.util.TreeMap;
+import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
+import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricMaterialBuilder;
 import net.fabricmc.fabric.api.tool.attribute.v1.FabricToolTags;
@@ -35,6 +35,8 @@ import net.minecraft.block.MaterialColor;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.BuiltinRegistries;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.util.registry.RegistryKey;
+import net.minecraft.world.gen.GenerationStep;
 import net.minecraft.world.gen.decorator.Decorator;
 import net.minecraft.world.gen.decorator.RangeDecoratorConfig;
 import net.minecraft.world.gen.feature.ConfiguredFeature;
@@ -42,41 +44,44 @@ import net.minecraft.world.gen.feature.Feature;
 import net.minecraft.world.gen.feature.OreFeatureConfig;
 
 public class MIMaterialSetup {
-    public static final Map<Identifier, ConfiguredFeature<?, ?>> ORE_GENERATORS = new TreeMap<>();
     // Materials
     public static final Material METAL_MATERIAL = new FabricMaterialBuilder(MaterialColor.IRON).build();
     public static final Material STONE_MATERIAL = new FabricMaterialBuilder(MaterialColor.STONE).build();
 
-    static {
+    public static void register() {
         MIMaterial fetcher = MIMaterials.gold; // force static load TODO : REFACTOR
 
         for (MIMaterial material : MIMaterial.getAllMaterials()) {
 
-            for (String block_type : material.getBlockType()) {
+            for (String blockType : material.getBlockType()) {
 
-                Block block = null;
-                if (!block_type.equals("ore")) {
+                Block block;
+                if (!blockType.equals("ore")) {
                     block = new MaterialBlock(FabricBlockSettings.of(METAL_MATERIAL).hardness(5.0f).resistance(6.0f)
                             .breakByTool(FabricToolTags.PICKAXES, 0).requiresTool(), material.getId(), "block");
                 } else {
                     block = new MaterialBlock(FabricBlockSettings.of(STONE_MATERIAL).hardness(3.0f).resistance(3.0f)
                             .breakByTool(FabricToolTags.PICKAXES, 1).requiresTool(), material.getId(), "ore");
                 }
-                material.saveBlock(block_type, block);
-
+                material.saveBlock(blockType, block);
+                Identifier identifier = new MIIdentifier(material.getId() + "_" + blockType);
+                Registry.register(Registry.BLOCK, identifier, material.getBlock(blockType));
             }
 
             if (material.hasOre()) {
-                ConfiguredFeature<?, ?> ore_generator = Feature.ORE
-                        .configure(new OreFeatureConfig(OreFeatureConfig.Rules.BASE_STONE_OVERWORLD, material.getBlock("ore").getDefaultState(),
-                                material.getVeinsSize())) // vein size
+                Block block = material.getBlock("ore");
+                ConfiguredFeature<?, ?> oreGenerator = Feature.ORE
+                        .configure(
+                                new OreFeatureConfig(OreFeatureConfig.Rules.BASE_STONE_OVERWORLD, block.getDefaultState(), material.getVeinsSize())) // vein
+                                                                                                                                                     // size
                         .decorate(Decorator.RANGE.configure(new RangeDecoratorConfig(0, // bottom offset
                                 0, // min y level
                                 material.getMaxYLevel()))) // max y level
                         .spreadHorizontally().repeat(material.getVeinsPerChunk()); // number of veins per chunk
-                Identifier oregen_id = new MIIdentifier("ore_generator_" + material.getId());
-                ORE_GENERATORS.put(oregen_id, ore_generator);
-                Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, oregen_id, ore_generator);
+                Identifier oregenId = new MIIdentifier("ore_generator_" + material.getId());
+                Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, oregenId, oreGenerator);
+                RegistryKey<ConfiguredFeature<?, ?>> featureKey = RegistryKey.of(Registry.CONFIGURED_FEATURE_WORLDGEN, oregenId);
+                BiomeModifications.addFeature(BiomeSelectors.foundInOverworld(), GenerationStep.Feature.UNDERGROUND_ORES, featureKey);
             }
         }
     }